docker/ml-service/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
python-multipart